Rustによる簡易的なWebサーバを実装
